polyval: Add runtime PCLMULQDQ detection
When the `std` feature is enabled (which it is now by default), this
adds runtime detection for PCLMULQDQ support on x86/x86_64
architectures.

The detection happens once at the time `Polyval` is instantated. The
`polyval::field::Element` type has been changed into an enum which
remembers the detection result, and its API changed to operate on
bytestring representations of POLYVAL field elements.

This appears to have a negligable performance impact.