Use correct `pshufd` intrinsic
Previously the implementation used the `_mm_shuffle_pd` intrinsic which
maps to the `shufpd` (floating point shuffle) instruction.

This still produced the correct assembly, as LLVM noticed the inputs and
outputs were integers and substituted `pshufd` for `shufpd`, which is
the correct ASM according to this QuarksLab blog post:

https://blog.quarkslab.com/reversing-a-finite-field-multiplication-optimization.html

This changes the code to call the proper `pshufd` intrinsic directly
(`_mm_shuffle_epi32`), which is cleaner.