[WIP] poly1305: AVX2 based on Bhattacharyya/Sarkar paper
Mechanical translation originally based on Corrode.

The original source code seems to be incomplete, but still amenable to a
mechanical translation:

https://github.com/Sreyosi/Improved-SIMD-Implementation-of-Poly1305

Notably its internal buffering in the original source code is commented
out, but not part of the `UniversalHash`-trait API.

We could potentially throw away all of the optimizations for
larger-than-block-size buffers and the incomplete code for buffering
non-block-aligned messages to get a baseline AVX2-optimized
implementation of Poly1305.

That said, this code is failing the basic test vectors, probably owing
to its incompleteness, and it probably makes more sense to base an AVX2
implementation on well-tested code which may perform slower but is known
to work and be generally robust.