POLYVAL rework
The main notable change in this PR is a restructuring of the CLMUL
backend along with a new reduction which avoids use of CLMUL (versus
previously using Shay Gueron's fast Montgomery reduction).

Additionally, this commit simplifies the implementation by having every
backend impl the `UniversalHash` trait directly. While adding a small
amount of boilerplate, This makes it easier to implement
backend-specific optimizations and should also make the backends easier
for rustc to optimize.