polyval: fix `zeroize` feature on autodetect backend
Adds a `Drop` impl for the autodetect backend which zeroizes when the
`zeroize` feature is enabled.

Due to the need for union members to be `Copy` until we can leverage
`ManuallyDrop` on MSRV 1.49+, the `soft` backends still don't implement
it properly.

This commit also adds a TODO to fix that when we bump MSRV.