polyval: `mulx` feature (#107)
Adds a feature for doubling POLYVAL field elements ("multiply-by-x"),
a.k.a. `mulX_POLYVAL` as defined in RFC 8452 Appendix A.

The main practical use of this feature is implementing GHASH in terms of
POLYVAL, and indeed this functionality was previously defined in the
GHASH crate.

However, this function is one which acts on POLYVAL field elements, and
while it's ultimately an implementation detail of GHASH, since it's an
operation over POLYVAL field elements the `polyval` crate seems like a
more proper home.

Additionally adds test vectors for this operation similar to the
`DBL_TEST_VECTORS` which test a similar operation in the `k256` and
`p256` elliptic curve crates.