ghash: use `polyval::mulx` (#109)
The `ghash` crate is implemented in terms of POLYVAL using a technique
described in RFC 8452 Appendix A.

Part of this implementation is a multiply-by-x operation over POLYVAL
field elements (a.k.a. `mulX_POLYVAL` in RFC 8452 terminology), which
was previously defined in the `ghash` crate itself.

However, this function was just added to the `polyval` crate in #107, so
the `ghash` crate can now use that implementation.