polyval: ARMv8 PMULL backend
The ARMv8 Cryptography Extensions provide a CLMUL-like instruction
called PMULL which can be used for computing GHASH/POLYVAL.

Additional background: https://eprint.iacr.org/2015/688.pdf

This commit adds a nightly-only PMULL accelerated backend for POLYVAL
based on this public domain C intrinsics implementation:

https://github.com/noloader/AES-Intrinsics/blob/master/clmul-arm.c

Adapting it to POLYVAL required changes similar to the CLMUL backend,
namely adapting the mask to use POLYVAL's polynomial (which is the
reverse of GHASH), and some additional work in the reduction to make it
"desrever" (as Shay Gueron likes to say).

Performance seems suboptimal, but still significantly better than the
software implementation by an order of magnitude. It seems that ARMv8
CPUs support a number of instruction fusions with PMULL, e.g.
`fuse-crypto-eor`, and we should investigate those.

Additionally it seems like we could better schedule operations on
multiple blocks in parallel.