polyval: pin `nightly` version used for ARMv8 testing (#151)
The build is currently failing:

https://github.com/RustCrypto/universal-hashes/runs/5629586918?check_suite_focus=true

It's for two reasons:

1. `aarch64_target_feature` was recently stabilized in
   rust-lang/rust#90621 but we still include it
2. There's a bug releating to `neon`/`fp` activation. See
   rust-lang/rust#91608 and rust-lang/rust#95044

Until one of the fixes in the second issue is merged, we can't really
make progress. So this commit pins to `nightly-2022-03-01` so we can
continue to have a clean build.

Once either of the above solutions lands we can remove
`aarch64_target_feature` and unpin nightly again.