polyval: impl `Reset`
This is useful in the implementation of `aes-gcm-siv`, as it allows for
calling the newly re-added `finalize_reset` method, which is useful to
avoid consuming `self` when computing the tag.

SIV mode encryption requires first computing the tag as a pass of
POLYVAL over the plaintext, where the tag also functions as a synthetic
initialization vector. With AES-GCM-SIV specifically, it also needs
access to the original nonce, which is XORed into the tag.

Consuming `self` means it's not possible to store the cipher, POLYVAL,
and nonce in the same struct, since after computing the SIV tag it needs
to be passed to the cipher.