Replace `armv8`/`force-soft` features with `cfg` attributes (#159)
Following the changes in e.g. the `aes` crate, this replaces cargo
features for this functionality with attributes which can be provided in
the toplevel build configuration as arguments to rustc:

- poly1305: `poly1305_force_soft`
- polyval: `polyval_armv8`, `polyval_force_soft`

This follows a similar naming scheme in the `aes` crate:
`aes_force_soft`

It would probably make sense to consolidate these so we have the same
names across crates like `aes` and `polyval` so it could be activated
with a single attribute, e.g. `rustcrypto_armv8`/`rustcrypto_force_soft`

However, for consistency this uses a similar scheme to what is already
released in the `aes` crate.